refactor(deps): relocate server-identity helpers to byos (decouple PR-B slice 3)
Slice 3 (final) of relocating the streamDeps helpers to internal/ (so a future
bintrail-console binary can reach them). Moves the server-identity pair out of
cmd/bintrail (package main) into internal/byos — reuniting them with the
SourceIdentity type they already operated on (byos/split.go):

- loadSourceIdentity    → byos.LoadSourceIdentity
- resolveServerIdentity → byos.ResolveServerIdentity

cmd/bintrail/serverutil.go and serverutil_test.go are deleted (git tracks both
as renames into internal/byos/identity.go + identity_test.go).

Pure relocation, zero behavior change: bodies moved verbatim and exported; the
only edits are package main→byos, the export-casing of the two names, and
dropping the byos. qualifier off SourceIdentity (now in-package). Callers
(snapshot/index/agent + stream's streamDeps) rewired to byos.*; the stale
`loadSourceIdentity` comment in agent.go updated.

Import edges added: byos→config (ParseSourceDSN), byos→serverid (ResolveServer).
Both safe — config and serverid do NOT import byos (grep-verified), so the
package graph stays acyclic. internal/streamrun untouched.

With this slice every streamDeps helper now lives in internal/. (Still TODO
before the PR-D supervisor move: doctor/provisioning/deriveServerID → internal/.)

Verified: go build/vet (unit + integration) clean; full unit suite green (incl.
the three TestLoadSourceIdentity_* relocated to internal/byos); monitor/
supervisor integration tests pass against MySQL — they exercise streamDeps →
streamrun.One running byos.ResolveServerIdentity end-to-end.
